+++ packages/yoroi-extension/app/components/wallet/send/WalletSendFormSteps/AssetsDropdown.scss
